Remove HashMap and HashSets from the code.
In some cases the code was using HashMaps when a std feature was enabled
which could lead to construction of a non-deterministic wasm bytes where
order is random.

In other cases where ordering did not matter, HashMap was still used
conditionally, which is not a direct replacement to a BTreeMap due to
performance implications etc.

Co-authored-by: George Pisaltu <georgep@casperlabs.io>